requirements.txt
flask==1.0.2
requests==2.18.4
dialogflow==0.4.0
python-dotenv==0.8.2
python_telegram_bot==12.7.0